[ONEM-32355] Implement DsMgr lib debug callbacks
Debug callbacks has been registered for libds and libds-rpc in order
to get debugs from these libs through the same debug interface that is used
by the process linking them.
Thanks to that temporal order of debugs is preserved and verbosity of
debugs can be controlled. If no debug callbacks are registered both libraries
are printing all their debug output to stdout.